pyproject.toml
[project]
name = "llm-psi-spike-predict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "groq>=1.0.0",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
]
